Split LPA step definitions out to separate context file.
